fixes #16 - moved compression to cross cutting
