+++ CrossCutting/CrossCutting.csproj
@@ -38,6 +38,7 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -45,6 +46,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CompressionDelegateHandler.cs" />
     <Compile Include="DisposableObject.cs" />
     <Compile Include="Enumerations\AddressLocation.cs" />
     <Compile Include="Extensions\TypeExtensions.cs" />
